rename to goroutine-explore
The expression language is going to be completely rewritten and there will be
substantial changes to the goroutine parser and overall UX. In order to avoid
confusion with the upstream project, rename to goroutine-explore.

Contributions for goroutine-explore will be under the Blue Oak Model
License (1.0.0), a permissive license with patent protections, remediations, and
just generally better drafting. Add license headers to all source files.