language description of a new evaluator
The existing evaluator uses `Knetic/govaluate` which is serviceable but uses it
to evaluate strings embedded inside Go syntax that the user is typing into the
REPL. This makes for awkward UX where you have to remember to single quote
strings within the expression. Having the rest of the code use the `go/ast`
makes it hard to improve features that could use user-supplied filters. And the
original library is no longer maintained.

The cheapest approach here would be to migrate to the maintained fork
`casbin/govaluate`. I also explored the option to migrate to the excellent
`expr` library, which is almost certainly be the most reasonable thing to
do. But it would be a lot more *fun* to implement our own expression language
and interpreter, so that's what we're going to do here.

The new evaluator will be built off to the side so that we don't have to worry
about how it's going to get wired up to the REPL until everything is closer to
being ready to use. This commit just includes a LANG.md file that we'll use to
replace the root README when we're done.

Ref: https://github.com/Knetic/govaluate/blob/master/ARCHIVED.md
Ref: https://github.com/expr-lang/expr
Ref: https://github.com/casbin/govaluate