compiler: initial implementation
First pass at our byte code compiler. Consumes tokens from the tokenizer in a
single pass and emits a chunk of byte code to be consumed by the VM. This
changeset includes implementation of where/delete filters, parenthetical
expressions, and pipelines. None of the emitted byte code is even minimally
optimized at this point (ex. no short-circuiting on boolean comparisons).

Right now there are two major unsatisfying design decisions we'll want to
revisit as we implement and refine the VM:

* The command/function parsing can't make up its mind on where to fan in from
  tokens to VM instructions. For some we have a slice of lexemes which we'll need
  to map to a slice of functions in the VM. Whereas parsing where/delete tokens
  emit a small pile of instructions because of the required control flow.

* Pipeline expressions are implemented as "suffix parsers" that just end the
  expression being parsed. This seems to work but the precedence values we end up
  with seem arbitrary.