vm: initial implementation
First pass at the byte code VM for the evaluator. It's mostly a stack-based VM
but with a couple of special-purpose registers for a scratch space for the
goroutine dump being built by the current filter expression and for the
goroutine currently being filtered.

None of the side-effect commands have implementations yet, but where/delete
filter functions and the related operations on goroutine fields are more-or-less
working.